Allow unlimited message sizes
While this is not globally considered good practice,
it will be required as the heightened demand of data
transmissions within Workiva grows with time.  This
commit is only a general idea and will require further
edits.  This is simply to get the conversation started.

This logic has been duplicated from the thrift-nats
repository.  Currently there are a few problems with
missing the last packet for large transmissions.

Source:
https://github.com/Workiva/thrift-nats/blob/7bf7c48/thrift_nats/nats_transport.go#L117-L123